+++ src/iso/builder.rs
@@ -12,7 +12,7 @@ use crate::iso::boot_catalog::{
 use crate::iso::dir_record::IsoDirEntry;
 use crate::iso::volume_descriptor::{update_total_sectors_in_pvd, write_volume_descriptors};
 use crate::utils::{ISO_SECTOR_SIZE, pad_to_lba};
-// Import Mbr struct
+
 
 /// Represents a file within the ISO filesystem.
 pub struct IsoFile {
