+++ src/iso/gpt.rs
@@ -54,12 +54,7 @@ impl GptHeader {
             _reserved1: [0; 420],
         };
 
-        // Calculate header CRC32
-        let mut header_bytes: [u8; mem::size_of::<GptHeader>()] = unsafe { mem::transmute(header) };
-        header_bytes[16..20].copy_from_slice(&[0; 4]); // Zero out header_crc32 field for calculation
-        let mut hasher = Hasher::new();
-        hasher.update(&header_bytes);
-        header.header_crc32 = hasher.finalize();
+        
 
         header
     }
@@ -91,9 +86,13 @@ impl GptPartitionEntry {
         ending_lba: u64,
         partition_name: &str,
     ) -> Self {
-        let partition_type_guid_bytes = Uuid::parse_str(partition_type_guid).unwrap().into_bytes();
+        let partition_type_guid_bytes = Uuid::parse_str(partition_type_guid)
+            .expect("Failed to parse partition type GUID")
+            .into_bytes();
         let unique_partition_guid_bytes =
-            Uuid::parse_str(unique_partition_guid).unwrap().into_bytes();
+            Uuid::parse_str(unique_partition_guid)
+                .expect("Failed to parse unique partition GUID")
+                .into_bytes();
 
         let mut name_bytes = [0u16; 36];
         for (i, c) in partition_name.encode_utf16().take(36).enumerate() {
@@ -204,70 +203,4 @@ pub fn write_gpt_structures<W: Write + Seek>(
     Ok(())
 }
 
-// MBR related functions (from mbr.rs, adapted for GPT hybrid)
-#[repr(C, packed)]
-#[derive(Debug, Clone, Copy, Default)]
-pub struct MbrPartitionEntry {
-    pub bootable: u8,
-    pub starting_chs: [u8; 3],
-    pub partition_type: u8,
-    pub ending_chs: [u8; 3],
-    pub starting_lba: u32,
-    pub size_in_lba: u32,
-}
 
-#[repr(C, packed)]
-#[derive(Debug, Clone, Copy)]
-pub struct Mbr {
-    pub boot_code: [u8; 440],
-    pub disk_signature: u32,
-    pub reserved: u16,
-    pub partition_table: [MbrPartitionEntry; 4],
-    pub boot_signature: u16,
-}
-
-impl Default for Mbr {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Mbr {
-    pub fn new() -> Self {
-        Mbr {
-            boot_code: [0; 440],
-            disk_signature: 0,
-            reserved: 0,
-            partition_table: [MbrPartitionEntry::default(); 4],
-            boot_signature: 0xAA55,
-        }
-    }
-
-    pub fn write_to<W: Write + Seek>(&self, writer: &mut W) -> io::Result<()> {
-        let bytes: [u8; mem::size_of::<Mbr>()] = unsafe { mem::transmute(*self) };
-        writer.write_all(&bytes)?;
-        Ok(())
-    }
-}
-
-pub fn create_mbr_for_gpt_hybrid(total_lbas: u32, is_isohybrid: bool) -> io::Result<Mbr> {
-    let mut mbr = Mbr::new();
-
-    if is_isohybrid {
-        // Protective MBR for GPT
-        mbr.partition_table[0].bootable = 0x00; // Not bootable
-        mbr.partition_table[0].partition_type = 0xEE; // GPT Protective MBR
-        mbr.partition_table[0].starting_lba = 1; // Starts after MBR itself
-        mbr.partition_table[0].size_in_lba = total_lbas - 1; // Spans rest of the disk
-    } else {
-        // Standard MBR for El Torito (if not isohybrid)
-        // This part might need more specific logic depending on your El Torito implementation
-        // For now, a simple bootable partition covering the whole disk (excluding MBR)
-        mbr.partition_table[0].bootable = 0x80; // Bootable
-        mbr.partition_table[0].partition_type = 0xEF; // EFI System Partition (placeholder, adjust as needed)
-        mbr.partition_table[0].starting_lba = 1; // Starts after MBR itself
-        mbr.partition_table[0].size_in_lba = total_lbas - 1; // Spans rest of the disk
-    }
-
-    Ok(mbr)
-}
